build: applied types to index.js (#3816)
* build: added basic type checking

this is just the very basic configuration. As I discovered some bugs I wanted these to be fixed first before the next iteration

* format: apply format

* build: applied types to index.js

This is just the first of multiple iterations. I tried to keep the size of the PR reasonable.

* build: fixed some wrong types from the prev commit

* format: applied format

* build: first implementation of the Activity type

* fix(custom.d.ts): fixed some typos

* fix: resolve minor issues from code review

* refactor: validateActivity now accepts Partial<Activity>

removes the need for explicit casts in broker implementations

Co-authored-by: Tobias <ikosaeder@protonmail.com>